# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Backend API",
    "version": "1.0.0",
    "description": "API documentation cho hệ thống Learning Management System (LMS)",
    "contact": {
      "name": "LMS Team",
      "email": "admin@lms.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.lms.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập JWT token để xác thực"
      }
    },
    "schemas": {
...
      "ForumPost": {
        "type": "object",
        "required": [
          "title",
          "content",
          "authorId",
          "classId"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID của bài viết"
          },
          "title": {
            "type": "string",
            "minLength": 5,
            "maxLength": 200,
            "description": "Tiêu đề bài viết"
          },
          "content": {
            "type": "string",
            "minLength": 10,
            "maxLength": 10000,
            "description": "Nội dung bài viết"
          },
          "authorId": {
            "type": "integer",
            "description": "ID tác giả"
          },
          "classId": {
            "type": "integer",
            "description": "ID lớp học"
          },
          "isPinned": {
            "type": "boolean",
            "description": "Bài viết được ghim"
          },
          "isLocked": {
            "type": "boolean",
            "description": "Bài viết bị khóa"
          },
          "viewCount": {
            "type": "integer",
            "description": "Số lượt xem"
          },
          "likeCount": {
            "type": "integer",
            "description": "Số lượt thích"
          },
          "commentCount": {
            "type": "integer",
            "description": "Số bình luận"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Danh sách tags"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "File đính kèm"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
...
  },
...
  "paths": {
...
    "/api/v1/forum/classes/{classId}/posts": {
      "get": {
        "summary": "Lấy danh sách bài viết theo lớp",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "classId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID lớp học"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Trang hiện tại"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Số bài viết mỗi trang"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "recent",
                "popular",
                "discussed"
              ],
              "default": "recent"
            },
            "description": "Cách sắp xếp"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách bài viết thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "posts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ForumPost"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forum/posts": {
      "post": {
        "summary": "Tạo bài viết mới",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "content",
                  "classId"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 10000
                  },
                  "classId": {
                    "type": "integer"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo bài viết thành công"
          }
        }
      }
    },
    "/api/v1/forum/posts/{id}": {
      "get": {
        "summary": "Lấy chi tiết bài viết",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID bài viết"
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy bài viết thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ForumPost"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Cập nhật bài viết",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 5,
                    "maxLength": 200
                  },
                  "content": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 10000
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "attachments": {
                    "type": "array"
                  },
                  "isPinned": {
                    "type": "boolean"
                  },
                  "isLocked": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật bài viết thành công"
          }
        }
      },
      "delete": {
        "summary": "Xóa bài viết",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa bài viết thành công"
          }
        }
      }
    },
    "/api/v1/forum/posts/{postId}/comments": {
      "get": {
        "summary": "Lấy bình luận của bài viết",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "postId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy bình luận thành công"
          }
        }
      }
    },
    "/api/v1/forum/comments": {
      "post": {
        "summary": "Tạo bình luận mới",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "postId",
                  "content"
                ],
                "properties": {
                  "postId": {
                    "type": "integer"
                  },
                  "content": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "parentId": {
                    "type": "integer"
                  },
                  "attachments": {
                    "type": "array"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo bình luận thành công"
          }
        }
      }
    },
    "/api/v1/forum/comments/{id}": {
      "put": {
        "summary": "Cập nhật bình luận",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "content"
                ],
                "properties": {
                  "content": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "attachments": {
                    "type": "array"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật bình luận thành công"
          }
        }
      },
      "delete": {
        "summary": "Xóa bình luận",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa bình luận thành công"
          }
        }
      }
    },
    "/api/v1/forum/likes": {
      "post": {
        "summary": "Toggle like/unlike",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetType",
                  "targetId"
                ],
                "properties": {
                  "targetType": {
                    "type": "string",
                    "enum": [
                      "post",
                      "comment"
                    ]
                  },
                  "targetId": {
                    "type": "integer"
                  },
                  "likeType": {
                    "type": "string",
                    "enum": [
                      "like",
                      "love",
                      "laugh",
                      "wow",
                      "sad",
                      "angry"
                    ],
                    "default": "like"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Toggle like thành công"
          }
        }
      }
    },
    "/api/v1/forum/likes/{targetType}/{targetId}": {
      "get": {
        "summary": "Lấy danh sách likes",
        "tags": [
          "Forum"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "targetType",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "post",
                "comment"
              ]
            }
          },
          {
            "in": "path",
            "name": "targetId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách likes thành công"
          }
        }
      }
    },
...
  }
}